# easyCasbin/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /easyCasbin/api/v1/user:
        put:
            tags:
                - User
            description: 更新用户
            operationId: User_UpdateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.UpdateUserInfo'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        post:
            tags:
                - User
            description: 创建用户
            operationId: User_CreateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.CreateUserInfo'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.UserInfoResponse'
    /easyCasbin/api/v1/user/login:
        post:
            tags:
                - User
            description: 登录
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.LoginRpl'
    /easyCasbin/api/v1/user/register:
        post:
            tags:
                - User
            description: 注册
            operationId: User_RegisterUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /easyCasbin/api/v1/user/{id}:
        get:
            tags:
                - User
            description: 通过 Id 查询用户
            operationId: User_GetUserById
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.UserInfoResponse'
    /easyCasbin/api/v1/user/{mobile}:
        get:
            tags:
                - User
            description: 通过 mobile 查询用户
            operationId: User_GetUserByMobile
            parameters:
                - name: mobile
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.UserInfoResponse'
    /easyCasbin/api/v1/users:
        get:
            tags:
                - User
            description: 获取用户列表
            operationId: User_GetUserList
            parameters:
                - name: pn
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.UserListResponse'
...
components:
    schemas:
...
        api.user.v1.CreateUserInfo:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                mobile:
                    type: string
            description: 创建用户
        api.user.v1.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
            description: 登录
        api.user.v1.LoginRpl:
            type: object
            properties:
                User:
                    $ref: '#/components/schemas/api.user.v1.UserInfoResponse'
                Token:
                    type: string
                ExpiresAt:
                    type: string
        api.user.v1.RegisterRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                mobile:
                    type: string
            description: 注册用户
        api.user.v1.UpdateUserInfo:
            type: object
            properties:
                id:
                    type: string
                nickName:
                    type: string
                gender:
                    type: string
                birthday:
                    type: string
        api.user.v1.UserInfoResponse:
            type: object
            properties:
                id:
                    type: string
                mobile:
                    type: string
                username:
                    type: string
            description: 用户信息
        api.user.v1.UserListResponse:
            type: object
            properties:
                total:
                    type: integer
                    format: int32
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.user.v1.UserInfoResponse'
            description: 用户列表
...